+++ client/src/pages/Login/SignUp.jsx
@@ -67,7 +67,11 @@ const SignUp = () => {
     </div>
    </section>
    <section className="hidden w-[60%] xl:block">
-    <img src="/bg.png" alt="image dashboard" className="h-full w-full"></img>
+    <img
+     src="/layout.png"
+     alt="image dashboard"
+     className="h-full w-full"
+    ></img>
    </section>
   </main>
  );
